+++ app/js/components/button/button.js
@@ -1,0 +1,17 @@
+import './btn-shownews.scss';
+
+const body = document.querySelector('body');
+
+const button = document.createElement('button');
+
+button.textContent = 'Show News';
+button.className = 'btn btn-danger btn-shownews';
+button.addEventListener('click', loadNews);
+body.appendChild(button);
+
+async function loadNews() {
+  await import('../../common');
+
+  const parent = button.parentElement;
+  parent.removeChild(button);
+}
