+++ app/js/index.js
@@ -1,21 +1,7 @@
 import 'bootstrap/scss/bootstrap.scss';
-import '../scss/btn-shownews.scss';
+
+import './components/button/button';
 
 //import './test.json';
 
-const body = document.querySelector('body');
-
-const button = document.createElement('button');
-
-button.textContent = 'Show News';
-button.className = 'btn btn-danger btn-shownews';
-button.addEventListener('click', loadNews);
-body.appendChild(button);
-
-async function loadNews() {
-  await import('./common');
-
-  const parent = button.parentElement;
-  parent.removeChild(button);
-}
